+++ packages/analytics/analytics-geo-map/CHANGELOG.md
@@ -3,6 +3,25 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.4.20](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/analytics-geo-map@1.4.19...@kong-ui-public/analytics-geo-map@1.4.20) (2025-10-27)
+
+**Note:** Version bump only for package @kong-ui-public/analytics-geo-map
+
+
+
+
+
+## [1.4.19](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/analytics-geo-map@1.4.18...@kong-ui-public/analytics-geo-map@1.4.19) (2025-10-27)
+
+
+### Bug Fixes
+
+* **deps:** update non-major [@kong](https://github.com/kong) dependencies with stable versions ([#2593](https://github.com/Kong/public-ui-components/issues/2593)) ([1283fdd](https://github.com/Kong/public-ui-components/commit/1283fdd7e9f6bccf0fbb5df160917e53d1293a03))
+
+
+
+
+
 ## [1.4.18](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/analytics-geo-map@1.4.17...@kong-ui-public/analytics-geo-map@1.4.18) (2025-10-24)