+++ packages/entities/entities-routes/CHANGELOG.md
@@ -3,6 +3,33 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.16.17](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-routes@3.16.16...@kong-ui-public/entities-routes@3.16.17) (2025-10-28)
+
+**Note:** Version bump only for package @kong-ui-public/entities-routes
+
+
+
+
+
+## [3.16.16](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-routes@3.16.15...@kong-ui-public/entities-routes@3.16.16) (2025-10-27)
+
+**Note:** Version bump only for package @kong-ui-public/entities-routes
+
+
+
+
+
+## [3.16.15](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-routes@3.16.14...@kong-ui-public/entities-routes@3.16.15) (2025-10-27)
+
+
+### Bug Fixes
+
+* **deps:** update non-major [@kong](https://github.com/kong) dependencies with stable versions ([#2593](https://github.com/Kong/public-ui-components/issues/2593)) ([1283fdd](https://github.com/Kong/public-ui-components/commit/1283fdd7e9f6bccf0fbb5df160917e53d1293a03))
+
+
+
+
+
 ## [3.16.14](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-routes@3.16.13...@kong-ui-public/entities-routes@3.16.14) (2025-10-21)
 
 **Note:** Version bump only for package @kong-ui-public/entities-routes